Moved test classes under org.scalatest.fixture in funsuite-test to org.scalatest.funsuite package.
