+++ jvm/funsuite-test/src/test/scala/org/scalatest/funsuite/FixtureAsyncFunSuiteLikeSpec2.scala
@@ -13,27 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.funsuite
 
-import scala.concurrent.{Promise, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest._
-import SharedHelpers.{EventRecordingReporter, thisLineNumber}
+import SharedHelpers.EventRecordingReporter
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
-import org.scalatest.exceptions.DuplicateTestNameException
 
 import scala.util.Success
 import org.scalatest
+import org.scalatest.funsuite
 
-class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
+class FixtureAsyncFunSuiteLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
-  describe("AsyncFunSuite") {
+  describe("AsyncFunSuiteLike") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuite with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -77,27 +75,26 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "test 5")
+      }
     }
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuite with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -131,29 +128,28 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "test 5")
+      }
     }
 
     it("should run tests that return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuite {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -164,7 +160,7 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
             SleepHelper.sleep(30)
             assert(count == 0)
             count = 1
-            succeed
+            Succeeded
           }
         }
 
@@ -173,7 +169,7 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
             assert(count == 1)
             SleepHelper.sleep(50)
             count = 2
-            succeed
+            Succeeded
           }
         }
 
@@ -188,22 +184,19 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     it("should run tests that does not return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuite {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -213,14 +206,14 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
           SleepHelper.sleep(30)
           assert(count == 0)
           count = 1
-          succeed
+          Succeeded
         }
 
         test("test 2") { fixture =>
           assert(count == 1)
           SleepHelper.sleep(50)
           count = 2
-          succeed
+          Succeeded
         }
 
         test("test 3") { fixture =>
@@ -232,13 +225,12 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     // SKIP-SCALATESTJS,NATIVE-START
@@ -249,7 +241,7 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -277,14 +269,17 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should run tests and its true async future in the same thread when use SerialExecutionContext") {
@@ -293,7 +288,7 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -341,19 +336,22 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -378,16 +376,18 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      status.waitUntilCompleted()
-      assert(!rep.testSucceededEventsReceived.isEmpty)
+
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(!rep.testSucceededEventsReceived.isEmpty)
+      }
     }
     // SKIP-SCALATESTJS,NATIVE-END
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -418,25 +418,24 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      }
     }
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -461,26 +460,23 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      }
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -491,21 +487,19 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
 
-      assert(infoList.size == 1)
-      assert(infoList(0).message == "hi there")
+        assert(infoList.size == 1)
+        assert(infoList(0).message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -517,27 +511,25 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -551,27 +543,25 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -582,21 +572,19 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
 
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -608,20 +596,18 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -635,20 +621,18 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -659,21 +643,19 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
 
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -685,20 +667,18 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -712,20 +692,18 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -736,21 +714,19 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
 
-      assert(markupList.size == 1)
-      assert(markupList(0).text == "hi there")
+        assert(markupList.size == 1)
+        assert(markupList(0).text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -762,27 +738,25 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuite  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -796,54 +770,25 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
-    }
-
-    it("should generate a DuplicateTestNameException when duplicate test name is detected") {
-      class TestSpec extends funsuite.FixtureAsyncFunSuite {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome = test("testing")
-        test("test 1") { fixture => succeed }
-        test("test 1") { fixture => succeed }
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
       }
-      val e = intercept[DuplicateTestNameException] {
-        new TestSpec
-      }
-      assert("AsyncFunSuiteSpec.scala" == e.failedCodeFileName.get)
-      assert(e.failedCodeLineNumber.get == thisLineNumber - 6)
-      assert(!e.cause.isDefined)
-    }
-
-    it("should generate a DuplicateTestNameException when duplicate test name is detected using ignore") {
-      class TestSpec extends funsuite.FixtureAsyncFunSuite {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome = test("testing")
-        test("test 1") { fixture => succeed }
-        ignore("test 1") { fixture => succeed }
-      }
-      val e = intercept[DuplicateTestNameException] {
-        new TestSpec
-      }
-      assert("AsyncFunSuiteSpec.scala" == e.failedCodeFileName.get)
-      assert(e.failedCodeLineNumber.get == thisLineNumber - 6)
-      assert(!e.cause.isDefined)
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends funsuite.FixtureAsyncFunSuite {
+      class TestSpec extends funsuite.FixtureAsyncFunSuiteLike {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
@@ -867,12 +812,12 @@ class AsyncFunSuiteSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new TestSpec
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(reporter.testStartingEventsReceived.length == 3)
-      assert(reporter.testSucceededEventsReceived.length == 3)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { r =>
+        assert(reporter.testStartingEventsReceived.length == 3)
+        assert(reporter.testSucceededEventsReceived.length == 3)
+      }
     }
 
   }
