+++ jvm/funsuite-test/src/test/scala/org/scalatest/funsuite/FixtureAsyncFunSuiteSpec2.scala
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.funsuite
 
-import scala.concurrent.{Promise, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest._
 import SharedHelpers.EventRecordingReporter
 import org.scalatest.concurrent.SleepHelper
@@ -23,17 +23,14 @@ import org.scalatest.events.{InfoProvided, MarkupProvided}
 
 import scala.util.Success
 import org.scalatest
-import org.scalatest.funsuite
 
-class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
+class FixtureAsyncFunSuiteSpec2 extends scalatest.funspec.AsyncFunSpec {
 
-  describe("AsyncFunSuiteLike") {
+  describe("AsyncFunSuite") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuite with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -77,27 +74,26 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "test 5")
+      }
     }
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuite with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -131,29 +127,28 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "test 5")
+      }
     }
 
     it("should run tests that return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -164,7 +159,7 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
             SleepHelper.sleep(30)
             assert(count == 0)
             count = 1
-            succeed
+            Succeeded
           }
         }
 
@@ -173,7 +168,7 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
             assert(count == 1)
             SleepHelper.sleep(50)
             count = 2
-            succeed
+            Succeeded
           }
         }
 
@@ -188,22 +183,19 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     it("should run tests that does not return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSuite extends funsuite.FixtureAsyncFunSuiteLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSuite extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -213,14 +205,14 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
           SleepHelper.sleep(30)
           assert(count == 0)
           count = 1
-          succeed
+          Succeeded
         }
 
         test("test 2") { fixture =>
           assert(count == 1)
           SleepHelper.sleep(50)
           count = 2
-          succeed
+          Succeeded
         }
 
         test("test 3") { fixture =>
@@ -232,13 +224,12 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     // SKIP-SCALATESTJS,NATIVE-START
@@ -249,7 +240,7 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -277,14 +268,17 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should run tests and its true async future in the same thread when use SerialExecutionContext") {
@@ -293,7 +287,7 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -341,19 +335,22 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -378,16 +375,18 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      status.waitUntilCompleted()
-      assert(!rep.testSucceededEventsReceived.isEmpty)
+
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(!rep.testSucceededEventsReceived.isEmpty)
+      }
     }
     // SKIP-SCALATESTJS,NATIVE-END
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -422,21 +421,23 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.waitUntilCompleted()
       // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      }
     }
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends funsuite.FixtureAsyncFunSuiteLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+      class ExampleSpec extends funsuite.FixtureAsyncFunSuite {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -465,21 +466,22 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.waitUntilCompleted()
       // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "test 3")
+      }
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -490,21 +492,19 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
 
-      assert(infoList.size == 1)
-      assert(infoList(0).message == "hi there")
+        assert(infoList.size == 1)
+        assert(infoList(0).message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -516,27 +516,25 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -550,27 +548,25 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -581,21 +577,19 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
 
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -607,20 +601,18 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -634,20 +626,18 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -658,21 +648,19 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
 
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -684,20 +672,18 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -711,20 +697,18 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -735,21 +719,19 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
 
-      assert(markupList.size == 1)
-      assert(markupList(0).text == "hi there")
+        assert(markupList.size == 1)
+        assert(markupList(0).text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -761,27 +743,25 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends funsuite.FixtureAsyncFunSuiteLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-
+      class MySuite extends funsuite.FixtureAsyncFunSuite  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -795,24 +775,25 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
+      }
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends funsuite.FixtureAsyncFunSuiteLike {
+      class TestSpec extends funsuite.FixtureAsyncFunSuite {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
@@ -836,12 +817,12 @@ class AsyncFunSuiteLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new TestSpec
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(reporter.testStartingEventsReceived.length == 3)
-      assert(reporter.testSucceededEventsReceived.length == 3)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { r =>
+        assert(reporter.testStartingEventsReceived.length == 3)
+        assert(reporter.testSucceededEventsReceived.length == 3)
+      }
     }
 
   }
